LayoutTests/requestidlecallback/requestidlecallback-in-page-cache-expected.txt

This tests that when requestIdleCallback is not enabled, requestIdleCallback and IdleDeadline are not defined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.persisted is true
PASS logsA.length is 0
PASS logsB.length is 0
PASS logsA.length is 4
PASS logsA.join(", ") is "A1, A2, A3, A4"
PASS logsB.length is 3
PASS logsB.join(", ") is "B1, B2, B3"
PASS successfullyParsed is true

TEST COMPLETE
